// src/client/app/model/modal.ts
export class Modal {
    firstName: String;
    lastName: String;
    password: String;
    dob: String;
    gender: String;
    email: String;
    phoneNo: String;
    goal: String;
    height: number;
    weight: number;
    pal: number;
    trainingPerWeek: String;
    dietPlan: String;
    planId: String;
    planName: String;
    targetWeight: number;
    carbs: String;
    protein: String;
    fat: String;
    calorie: String;
}